requirements.txt
pygame>=2.1.2
pyserial
python-escpos
opencv-python